Add rexml dependency
This gem uses rexml. When running the tests with bundler, this
dependency used to be pulled in by rubocop. However, rubocop 1.66 no
longer depends on rexml, so the tests now fail when running with
up-to-date dependencies.

When using the dnote tool, the absence of an explicit dependency was
never a problem since rexml is a bundled gem.

This change adds an explicit dependency on rexml.